move the hook into a default module
